.flake8: Remove unused config options (#7)
What started as just removing references to `apollo_gateway` became a
realization that the `.flake8` has a lot of stuff from MyJobs that
doesn't belong in this repo.

An argument could be made that the whole file ought to be deleted since
Flake8 isn't installed in this package, but I don't think there's harm
in keeping it around for now.